NPC and Player Update
NPC:
NPC can now choose whether or not to run to the next target. (simple)
Gives them the appearance of beeing more human like.
-NEED TO ADD: Points along path move on perlin noise, this makes their movement harder to predict and makes them seem more human like.

-Player
Seeker player has began to be implemented.
Not sure on what type of player it should be, sniper (cross hair) or physical (character).
For the moment: Sniper character is being worked on.